requirements.txt
tensorflow==2.15.0
transformers==4.40.0
sentencepiece>=0.1.99
numpy<2.0
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.13
Pillow>=10.0
tqdm>=4.65
